implemented researcher as a nested chat
